Update campground permissions and move middlewares
