+++ middleware.js
@@ -1,3 +1,7 @@
+const { campgroundSchema, reviewSchema } = require('./schemas');
+const Campground = require('./models/campground');
+const ExpressError = require('./utils/ExpressError');
+
 module.exports.isLoggedIn = (req, res, next) => {
   if (!req.isAuthenticated()) {
     req.session.returnTo = req.originalUrl;
@@ -10,6 +14,34 @@ module.exports.isLoggedIn = (req, res, next) => {
 module.exports.storeReturnTo = (req, res, next) => {
   if (req.session.returnTo) {
     res.locals.returnTo = req.session.returnTo;
+  }
+  next();
+};
+
+module.exports.isAuthor = async (req, res, next) => {
+  const { id } = req.params;
+  const campground = await Campground.findById(id);
+  if (!campground.author.equals(req.user._id)) {
+    req.flash('error', 'You do no have permission to do that.');
+    return res.redirect(`/campgrounds/${id}`);
+  }
+  next();
+};
+
+module.exports.validateCampground = (req, res, next) => {
+  const { error } = campgroundSchema.validate(req.body);
+  if (error) {
+    const message = error.details.map(detail => detail.message).join(',');
+    throw new ExpressError(message, 400);
+  }
+  next();
+};
+
+module.exports.validateReview = (req, res, next) => {
+  const { error } = reviewSchema.validate(req.body);
+  if (error) {
+    const message = error.details.map(detail => detail.message).join(',');
+    throw new ExpressError(message, 400);
   }
   next();
 };
