+++ routes/reviews.js
@@ -1,20 +1,9 @@
 const express = require('express');
 const router = express.Router({ mergeParams: true });
-const { reviewSchema } = require('../schemas');
+const { validateReview } = require('../middleware');
 const Campground = require('../models/campground');
 const Review = require('../models/review');
 const catchAsync = require('../utils/catchAsync');
-const ExpressError = require('../utils/ExpressError');
-
-const validateReview = (req, res, next) => {
-  const { error } = reviewSchema.validate(req.body);
-  if (error) {
-    const message = error.details.map(detail => detail.message).join(',');
-    throw new ExpressError(message, 400);
-  } else {
-    next();
-  }
-};
 
 router.post(
   '/',
